+++ photoview/src/main/java/com/github/chrisbanes/photoview/listener/OnOutsidePhotoTapListener.java
@@ -1,4 +1,4 @@
-package com.github.chrisbanes.photoview;
+package com.github.chrisbanes.photoview.listener;
 
 import android.widget.ImageView;
