+++ photoview/src/main/java/com/github/chrisbanes/photoview/listener/OnScaleChangedListener.java
@@ -1,4 +1,4 @@
-package com.github.chrisbanes.photoview;
+package com.github.chrisbanes.photoview.listener;
 
 
 /**
