+++ photoview/src/main/java/com/github/chrisbanes/photoview/listener/OnViewTapListener.java
@@ -1,4 +1,4 @@
-package com.github.chrisbanes.photoview;
+package com.github.chrisbanes.photoview.listener;
 
 import android.view.View;
